reducer consumes GeoJson; stations no longer rendering
