+++ src/components/SubwayStationsLayer.js
@@ -1,55 +1,8 @@
 import React from "react";
-import SubwayStations from '../data/SubwayStations.geojson.json'
-import SubwayStops from '../data/SubwayStops.json'
 import L from 'leaflet'
 import { GeoJSON, useMap } from 'react-leaflet'
 import { GlobalStationDispatchContext, GlobalStationStateContext } from "../context/GlobalContextProvider";
 import { loadStations, markStart, markEnd } from '../store/reducers/stations'
-
-// mapping from station name to station data
-const stationFromName = SubwayStops.reduce((accum, stop) => {
-  if (!stop.id.match(/\d+$/)) return accum
-  let stopName = stop.name
-  // put st/nd/rd/th after street numbers
-  for (const match of stopName.matchAll(/(\d+) /g)) {
-    let [orig, number] = match
-    const lastDigit = number % 10
-    if (lastDigit == 1) {
-      number = number + 'st'
-    } else if (lastDigit == 2) {
-      number = number + 'nd'
-    } else if (lastDigit == 3) {
-      number = number + 'rd'
-    } else {
-      number = number + 'th'
-    }
-    stopName = stopName.replaceAll(orig, `${number} `)
-  }
-  // convert Av -> Ave, Ft -> Fort
-  stopName = stopName
-    .replaceAll(' Av', ' Ave')
-    .replaceAll('-', ' - ')
-    .replaceAll('Ft ', 'Fort ')
-
-  accum[stopName] = stop
-  return accum
-}, {})
-
-// injects MTA station ID into each station feature
-const stationsWithId = { ...SubwayStations };
-stationsWithId.features = SubwayStations.features.map(f => {
-  const station = stationFromName[f.properties.name]
-  // console.log(station ? station : f.properties.name)
-  if (!station) return { wtf: true }
-  return {
-    ...f,
-    properties: {
-      ...f.properties,
-      id: station.id,
-    },
-    geometry: { ...f.geometry, coordinates: [...f.geometry.coordinates] }
-  }
-}).filter(s => s.properties && s.properties.id)
 
 function stationToMarker(station, latlng) {
   const markerStyle = {
@@ -62,6 +15,7 @@ function stationToMarker(station, latlng) {
   if (typeof window === 'undefined') {
     return null
   }
+  console.log(station)
   return new L.CircleMarker(latlng, markerStyle)
 }
 
@@ -96,7 +50,7 @@ const SubwayStationsLayer = () => {
   }, [map])
 
   return (
-    <GeoJSON data={stationsWithId} pointToLayer={stationToMarker}/>
+    <GeoJSON data={stationList} pointToLayer={stationToMarker}/>
   )
 }
